+++ src/common/constants/api.js
@@ -1,2 +1,3 @@
 export const INVENTORY_INSTANCE_API = 'inventory/instances';
 export const MARC_RECORD_API = 'records-editor/records';
+export const MARC_RECORD_STATUS_API = `${MARC_RECORD_API}/status`;